multiGPU_requirements.txt
# v.2021-08-09
absl-py==0.13.0
adamp==0.3.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
avro==1.11.0
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
elasticsearch==7.5.1
emoji==1.2.0
et-xmlfile==1.1.0
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.7.0
furl==2.1.2
future==0.18.2
gast==0.3.3
gensim==4.0.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
joblib==1.0.1
JPype1==1.3.0
jupyterthemes==0.20.0
Keras-Preprocessing==1.1.2
konlpy==0.5.2
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.5.1
multidict==5.1.0
numpy==1.19.5
lz4==4.0.0
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.0
pandas==1.3.1
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1    #If Windows:  psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
selenium==3.141.0
six==1.15.0
soynlp==0.0.493
sqlalchemy==1.4.22
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.10.3
# torch==1.9.0
torch-optimizer==0.3.0
tqdm==4.62.0
transformers==4.11.3
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.4
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3